pyproject.toml
[project]
name = "mqtt-system-monitor"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "paho-mqtt>=2.1.0",
    "psutil>=5.9.8",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"